+++ week4/src/constants/api.ts
@@ -1,9 +1,8 @@
 export const BASE_URL: string = import.meta.env.VITE_API_BASE_URL;
-export const ACCESS_TOKEN_KEY = 'ACCESS_TOKEN';
 
 export const END_POINT = {
   USER: '/user',
   LOGIN: '/login',
   MY_HOBBY: '/user/my-hobby',
   USER_HOBBY: (no: number) => `/user/${no}/hobby`,
-};
+} as const;
