+++ week4/src/constants/config.ts
@@ -1,1 +1,3 @@
+export const ACCESS_TOKEN_KEY = 'ACCESS_TOKEN';
+
 export const MAX_INPUT_LENGTH = 8;
